refactor(reservations): split fixed-slot template mutations
