+++ apps/server/src/domain/reservations/services/fixed-slot-template.service.ts
@@ -2,41 +2,21 @@ import { Context, Effect, Layer, Option } from "effect";
 
 import type { StationQueryRepo } from "@/domain/stations";
 
-import { makeBikeRepository } from "@/domain/bikes";
-import {
-  defectOn,
-  isWallClockWithinOvernightOperationsWindow,
-  makeFixedSlotTemplateStartOutsideOperatingHoursError,
-} from "@/domain/shared";
-import { makeStationQueryRepository, StationQueryRepository } from "@/domain/stations";
-import { Prisma } from "@/infrastructure/prisma";
-import { PrismaTransactionError, runPrismaTransaction } from "@/lib/effect/prisma-tx";
+import { StationQueryRepository } from "@/domain/stations";
 
-import type { ReservationCommandRepo } from "../repository/reservation-command.repository";
 import type { ReservationQueryRepo } from "../repository/reservation-query.repository";
 import type { FixedSlotTemplateService } from "./fixed-slot-template/fixed-slot-template.types";
 
 import {
-  FixedSlotTemplateCancelConflict,
-  FixedSlotTemplateConflict,
-  FixedSlotTemplateDateLocked,
-  FixedSlotTemplateDateNotFound,
-  FixedSlotTemplateDateNotFuture,
   FixedSlotTemplateNotFound,
-  FixedSlotTemplateStationNotFound,
 } from "../domain-errors";
-import { makeReservationCommandRepository, ReservationCommandRepository } from "../repository/reservation-command.repository";
-import { makeReservationQueryRepository, ReservationQueryRepository } from "../repository/reservation-query.repository";
+import { ReservationQueryRepository } from "../repository/reservation-query.repository";
+import { cancelFixedSlotTemplateForUser } from "./fixed-slot-template/fixed-slot-template.cancel";
+import { createFixedSlotTemplateForUser } from "./fixed-slot-template/fixed-slot-template.create";
 import {
-  applyTemplateMutation,
-  ensureTemplateMutationAllowed,
-} from "./fixed-slot-template/mutations";
-import {
-  normalizeSlotDate,
-  parseSlotDateKey,
-  parseSlotTimeValue,
-  toSlotDateKey,
-} from "./fixed-slot/fixed-slot.helpers";
+  removeFixedSlotTemplateDateForUser,
+  updateFixedSlotTemplateForUser,
+} from "./fixed-slot-template/fixed-slot-template.update";
 
 export type { FixedSlotTemplateService } from "./fixed-slot-template/fixed-slot-template.types";
 
@@ -52,69 +32,10 @@ export type { FixedSlotTemplateService } from "./fixed-slot-template/fixed-slot-
  */
 export function makeFixedSlotTemplateService(deps: {
   reservationQueryRepo: ReservationQueryRepo;
-  reservationCommandRepo: ReservationCommandRepo;
   stationRepo: StationQueryRepo;
 }): FixedSlotTemplateService {
   return {
-    createForUser: args =>
-      Effect.gen(function* () {
-        const { client } = yield* Prisma;
-        const now = args.now ?? new Date();
-        const today = normalizeSlotDate(now);
-        const slotStart = parseSlotTimeValue(args.slotStart);
-        const slotDates = args.slotDates.map(parseSlotDateKey);
-
-        if (isWallClockWithinOvernightOperationsWindow(slotStart)) {
-          return yield* Effect.fail(
-            makeFixedSlotTemplateStartOutsideOperatingHoursError(slotStart),
-          );
-        }
-
-        for (const slotDate of slotDates) {
-          if (slotDate.getTime() <= today.getTime()) {
-            return yield* Effect.fail(new FixedSlotTemplateDateNotFuture({
-              slotDate: toSlotDateKey(slotDate),
-            }));
-          }
-        }
-
-        return yield* runPrismaTransaction(client, tx =>
-          Effect.gen(function* () {
-            const txStationRepo = makeStationQueryRepository(tx);
-            const txReservationQueryRepo = makeReservationQueryRepository(tx);
-            const txReservationCommandRepo = makeReservationCommandRepository(tx);
-
-            const stationOpt = yield* txStationRepo.getById(args.stationId);
-            if (Option.isNone(stationOpt)) {
-              return yield* Effect.fail(new FixedSlotTemplateStationNotFound({
-                stationId: args.stationId,
-              }));
-            }
-
-            const conflictCount = yield* txReservationQueryRepo.countActiveFixedSlotTemplateConflicts(
-              args.userId,
-              slotStart,
-              slotDates,
-            );
-            // FIX: Enforce active fixed-slot overlap at DB level.
-            // This read-then-write check races under concurrent create/update requests and can still admit duplicate active templates.
-            if (conflictCount > 0) {
-              return yield* Effect.fail(new FixedSlotTemplateConflict({
-                userId: args.userId,
-                slotStart: args.slotStart,
-                slotDates: [...args.slotDates],
-              }));
-            }
-
-            return yield* txReservationCommandRepo.createFixedSlotTemplate({
-              userId: args.userId,
-              stationId: args.stationId,
-              slotStart,
-              slotDates,
-              updatedAt: now,
-            });
-          })).pipe(defectOn(PrismaTransactionError));
-      }),
+    createForUser: createFixedSlotTemplateForUser,
 
     listForUser: args =>
       deps.reservationQueryRepo.listFixedSlotTemplatesForUser(
@@ -144,190 +65,11 @@ export function makeFixedSlotTemplateService(deps: {
         return templateOpt.value;
       }),
 
-    cancelForUser: args =>
-      Effect.gen(function* () {
-        const { client } = yield* Prisma;
-        const now = args.now ?? new Date();
+    cancelForUser: cancelFixedSlotTemplateForUser,
 
-        return yield* runPrismaTransaction(client, tx =>
-          Effect.gen(function* () {
-            const txQueryRepo = makeReservationQueryRepository(tx);
-            const txCommandRepo = makeReservationCommandRepository(tx);
-            const bikeRepo = makeBikeRepository(tx);
+    updateForUser: updateFixedSlotTemplateForUser,
 
-            const templateOpt = yield* txQueryRepo.findFixedSlotTemplateByIdForUser(
-              args.userId,
-              args.templateId,
-            );
-            if (Option.isNone(templateOpt)) {
-              return yield* Effect.fail(new FixedSlotTemplateNotFound({
-                templateId: args.templateId,
-              }));
-            }
-
-            const pendingReservations = yield* txQueryRepo.listPendingFixedSlotReservationsByTemplateId(
-              args.templateId,
-            );
-
-            for (const reservation of pendingReservations) {
-              yield* txCommandRepo.updateStatus({
-                reservationId: reservation.id,
-                status: "CANCELLED",
-                updatedAt: now,
-              }).pipe(
-                Effect.catchTag("ReservationNotFound", () =>
-                  Effect.fail(new FixedSlotTemplateCancelConflict({ templateId: args.templateId }))),
-              );
-
-              if (reservation.bikeId) {
-                const released = yield* bikeRepo.releaseBikeIfReserved(reservation.bikeId, now);
-                if (!released) {
-                  return yield* Effect.fail(new FixedSlotTemplateCancelConflict({ templateId: args.templateId }));
-                }
-              }
-            }
-
-            if (templateOpt.value.status === "CANCELLED") {
-              return templateOpt.value;
-            }
-
-            return yield* txCommandRepo.updateFixedSlotTemplateStatus({
-              templateId: args.templateId,
-              status: "CANCELLED",
-              updatedAt: now,
-            });
-          }));
-      }).pipe(defectOn(PrismaTransactionError)),
-
-    updateForUser: args =>
-      Effect.gen(function* () {
-        const { client } = yield* Prisma;
-        const now = args.now ?? new Date();
-        const today = normalizeSlotDate(now);
-
-        return yield* runPrismaTransaction(client, tx =>
-          Effect.gen(function* () {
-            const txQueryRepo = makeReservationQueryRepository(tx);
-            const txCommandRepo = makeReservationCommandRepository(tx);
-            const bikeRepo = makeBikeRepository(tx);
-
-            const templateOpt = yield* txQueryRepo.findFixedSlotTemplateByIdForUser(
-              args.userId,
-              args.templateId,
-            );
-            if (Option.isNone(templateOpt)) {
-              return yield* Effect.fail(new FixedSlotTemplateNotFound({
-                templateId: args.templateId,
-              }));
-            }
-
-            const template = yield* ensureTemplateMutationAllowed(templateOpt.value, args.templateId);
-            const currentDateKeySet = new Set(template.slotDates.map(toSlotDateKey));
-            const lockedDateKeySet = new Set(
-              template.slotDates
-                .filter(slotDate => slotDate.getTime() <= today.getTime())
-                .map(toSlotDateKey),
-            );
-
-            const nextSlotStart = args.slotStart ? parseSlotTimeValue(args.slotStart) : template.slotStart;
-            if (
-              args.slotStart !== undefined
-              && isWallClockWithinOvernightOperationsWindow(nextSlotStart)
-            ) {
-              return yield* Effect.fail(
-                makeFixedSlotTemplateStartOutsideOperatingHoursError(nextSlotStart),
-              );
-            }
-            let nextSlotDates = template.slotDates;
-
-            if (args.slotDates) {
-              nextSlotDates = args.slotDates.map(parseSlotDateKey);
-              const nextDateKeySet = new Set(nextSlotDates.map(toSlotDateKey));
-
-              for (const lockedDateKey of lockedDateKeySet) {
-                if (!nextDateKeySet.has(lockedDateKey)) {
-                  return yield* Effect.fail(new FixedSlotTemplateDateLocked({
-                    slotDate: lockedDateKey,
-                  }));
-                }
-              }
-
-              for (const slotDate of nextSlotDates) {
-                const slotDateKey = toSlotDateKey(slotDate);
-                if (slotDate.getTime() <= today.getTime() && !currentDateKeySet.has(slotDateKey)) {
-                  return yield* Effect.fail(new FixedSlotTemplateDateNotFuture({ slotDate: slotDateKey }));
-                }
-              }
-            }
-
-            return yield* applyTemplateMutation({
-              userId: args.userId,
-              template,
-              templateId: args.templateId,
-              nextSlotStart,
-              nextSlotDates,
-              now,
-              tx,
-              txQueryRepo,
-              txCommandRepo,
-              bikeRepo,
-            });
-          }));
-      }).pipe(defectOn(PrismaTransactionError)),
-
-    removeDateForUser: args =>
-      Effect.gen(function* () {
-        const { client } = yield* Prisma;
-        const now = args.now ?? new Date();
-        const today = normalizeSlotDate(now);
-        const targetSlotDate = parseSlotDateKey(args.slotDate);
-        const targetSlotDateKey = toSlotDateKey(targetSlotDate);
-
-        return yield* runPrismaTransaction(client, tx =>
-          Effect.gen(function* () {
-            const txQueryRepo = makeReservationQueryRepository(tx);
-            const txCommandRepo = makeReservationCommandRepository(tx);
-            const bikeRepo = makeBikeRepository(tx);
-
-            const templateOpt = yield* txQueryRepo.findFixedSlotTemplateByIdForUser(
-              args.userId,
-              args.templateId,
-            );
-            if (Option.isNone(templateOpt)) {
-              return yield* Effect.fail(new FixedSlotTemplateNotFound({
-                templateId: args.templateId,
-              }));
-            }
-
-            const template = yield* ensureTemplateMutationAllowed(templateOpt.value, args.templateId);
-            const currentDateKeySet = new Set(template.slotDates.map(toSlotDateKey));
-            if (!currentDateKeySet.has(targetSlotDateKey)) {
-              return yield* Effect.fail(new FixedSlotTemplateDateNotFound({
-                templateId: args.templateId,
-                slotDate: targetSlotDateKey,
-              }));
-            }
-
-            if (targetSlotDate.getTime() <= today.getTime()) {
-              return yield* Effect.fail(new FixedSlotTemplateDateLocked({
-                slotDate: targetSlotDateKey,
-              }));
-            }
-
-            return yield* applyTemplateMutation({
-              userId: args.userId,
-              template,
-              templateId: args.templateId,
-              nextSlotStart: template.slotStart,
-              nextSlotDates: template.slotDates.filter(slotDate => toSlotDateKey(slotDate) !== targetSlotDateKey),
-              now,
-              tx,
-              txQueryRepo,
-              txCommandRepo,
-              bikeRepo,
-            });
-          }));
-      }).pipe(defectOn(PrismaTransactionError)),
+    removeDateForUser: removeFixedSlotTemplateDateForUser,
   };
 }
 
@@ -340,12 +82,10 @@ export const FixedSlotTemplateServiceLive = Layer.effect(
   FixedSlotTemplateServiceTag,
   Effect.gen(function* () {
     const reservationQueryRepo = yield* ReservationQueryRepository;
-    const reservationCommandRepo = yield* ReservationCommandRepository;
     const stationRepo = yield* StationQueryRepository;
 
     return makeFixedSlotTemplateService({
       reservationQueryRepo,
-      reservationCommandRepo,
       stationRepo,
     });
   }),
